django_app/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.11.2
channels==4.1.0
channels-redis==4.2.0
mongoengine==0.29.1
pymongo==4.8.0
PyJWT==2.9.0
bcrypt==4.2.0
requests==2.32.3
bleach==6.1.0
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0
daphne==4.1.2
python-dotenv==1.0.1
# Production dependencies
whitenoise==6.6.0
psycopg2-binary==2.9.9
# Logging and monitoring
python-json-logger==2.0.7
